+++ DesktopApplicationTemplate.Tests/ScrollBarStyleTests.cs
@@ -13,7 +13,6 @@ namespace DesktopApplicationTemplate.Tests
         [WpfFact]
         public void LightTheme_ScrollBarStyle_HasReducedWidth()
         {
-            Skip.IfNot(OperatingSystem.IsWindows(), "Requires Windows desktop runtime");
 
             Exception? ex = null;
             var thread = new Thread(() =>